Mod loading refactored into separate class. Preparatory work for basic dependency resolution to be implemented.
